packaging: @openibex/chain
The separate providers/index.d.ts has been merged into index.ts.
Rollup (used by tsup) gets confused when a separate index.d.ts is
present, as it only (!) considers exports in index.d.ts and ignores
index.ts completely.

Note the error is counter-intuitive as you'd start looking for errors in
src/providers/index.ts:
RollupError: "getChainProvider" is not exported by "src/providers/index.ts", imported by "src/index.ts".
