+++ packages/chain/README.md
@@ -1,0 +1,2 @@
+# @openibex/chain
+Description of the module TODO
